# api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "IoT 物联网平台后端 API（Go 重构版）",
        "title": "IoT Platform Go API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@iot-platform.local.local"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.7.0"
    },
    "host": "localhost:8182",
    "basePath": "/",
    "paths": {
        "/api/devices": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "查询用户设备列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/ent.Device"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    }
                },
                "security": [
                    {
                        "UserAuth": []
                    }
                ]
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "注册设备",
                "parameters": [
                    {
                        "description": "设备注册请求参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.DeviceParameters"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service.DeviceRegisterResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    }
                },
                "security": [
                    {
                        "UserAuth": []
                    }
                ]
            }
        },
        "/api/devices/{deviceId}": {
            "get": {
                "description": "设备元信息 + 在线状态 + sensors（定义+latest）/ actuators（定义）物模型数组",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "获取设备详情（物模型视图：定义 + 最近遥测 + 执行器）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设备ID（6位hex）",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    }
                },
                "security": [
                    {
                        "UserAuth": []
                    }
                ]
            }
        },
        "/api/devices/{deviceId}/actuators": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "查询设备执行器定义列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    }
                },
                "security": [
                    {
                        "UserAuth": []
                    }
                ]
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "创建执行器定义",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "执行器定义",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.ActuatorCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    }
                },
                "security": [
                    {
                        "UserAuth": []
                    }
                ]
            }
        },
        "/api/devices/{deviceId}/actuators/apply": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "下发执行器配置（编译进 DeviceConfig 并版本化下发）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    }
                },
                "security": [
                    {
                        "UserAuth": []
                    }
                ]
            }
        },
        "/api/devices/{deviceId}/actuators/{actuatorId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "查询单个执行器定义",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "执行器标识符",
                        "name": "actuatorId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    }
                },
                "security": [
                    {
                        "UserAuth": []
                    }
                ]
            }
        },
        "/api/devices/{deviceId}/actuators/{actuatorId}/delete": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "删除执行器定义",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "执行器标识符",
                        "name": "actuatorId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    }
                },
                "security": [
                    {
                        "UserAuth": []
                    }
                ]
            }
        },
        "/api/devices/{deviceId}/actuators/{actuatorId}/update": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "增量更新执行器定义",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "执行器标识符",
                        "name": "actuatorId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新字段（增量）",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.ActuatorUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    }
                },
                "security": [
                    {
                        "UserAuth": []
                    }
                ]
            }
        },
        "/api/devices/{deviceId}/commands": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device"
                ],
                "summary": "设备拉取待消费命令",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    }
                },
                "security": [
                    {
                        "DeviceAuth": []
                    }
                ]
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device"
                ],
                "summary": "向设备下发命令",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "命令内容",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.DownlinkCmdRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    }
                },
                "security": [
                    {
                        "UserAuth": []
                    }
                ]
            }
        },
        "/api/devices/{deviceId}/config": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "查询设备配置快照",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    }
                },
                "security": [
                    {
                        "UserAuth": []
                    },
                    {
                        "DeviceAuth": []
                    }
                ]
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "设置设备配置并下发",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "配置内容",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.DeviceConfigSaveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    }
                },
                "security": [
                    {
                        "UserAuth": []
                    }
                ]
            }
        },
        "/api/devices/{deviceId}/config/report": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "设备配置回执",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "实际生效配置与版本",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.DeviceConfigReport"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    }
                },
                "security": [
                    {
                        "DeviceAuth": []
                    }
                ]
            }
        },
        "/api/devices/{deviceId}/delete": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "删除设备",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    }
                },
                "security": [
                    {
                        "UserAuth": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
        "common.ApiResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "service.DeviceParameters": {
            "type": "object",
            "required": [
                "deviceName"
            ],
            "properties": {
                "deviceName": {
                    "type": "string"
                },
                "deviceType": {
                    "type": "string"
                },
                "firmwareVersion": {
                    "type": "string"
                },
                "ipAddress": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "macAddress": {
                    "type": "string"
                }
            }
        },
        "service.DeviceRegisterResponse": {
            "type": "object",
            "properties": {
                "deviceId": {
                    "type": "string"
                },
                "deviceToken": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "DeviceAuth": {
            "description": "设备 Token（设备注册后获取）",
            "type": "apiKey",
            "name": "X-Device-Token",
            "in": "header"
        },
        "UserAuth": {
            "description": "Bearer Token（用户 Sa-Token），登录后获取",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}